# paraphrasing-tool/app/generated/server/openapi_server/models/paraphrase_request.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class ParaphraseRequest(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, text=None):  # noqa: E501
        """ParaphraseRequest - a model defined in OpenAPI

        :param text: The text of this ParaphraseRequest.  # noqa: E501
        :type text: str
        """
        self.openapi_types = {
            'text': str
        }

        self.attribute_map = {
            'text': 'text'
        }

        self._text = text

    @classmethod
    def from_dict(cls, dikt) -> 'ParaphraseRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ParaphraseRequest of this ParaphraseRequest.  # noqa: E501
        :rtype: ParaphraseRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def text(self) -> str:
        """Gets the text of this ParaphraseRequest.


        :return: The text of this ParaphraseRequest.
        :rtype: str
        """
        return self._text

    @text.setter
    def text(self, text: str):
        """Sets the text of this ParaphraseRequest.


        :param text: The text of this ParaphraseRequest.
        :type text: str
        """
        if text is None:
            raise ValueError("Invalid value for `text`, must not be `None`")  # noqa: E501

        self._text = text