# paraphrasing-tool/app/generated/server/openapi_server/models/simple_choice.py
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model import Model
from openapi_server import util


class SimpleChoice(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, identifier=None, text=None):  # noqa: E501
        """SimpleChoice - a model defined in OpenAPI

        :param identifier: The identifier of this SimpleChoice.  # noqa: E501
        :type identifier: str
        :param text: The text of this SimpleChoice.  # noqa: E501
        :type text: str
        """
        self.openapi_types = {
            'identifier': str,
            'text': str
        }

        self.attribute_map = {
            'identifier': 'identifier',
            'text': 'text'
        }

        self._identifier = identifier
        self._text = text

    @classmethod
    def from_dict(cls, dikt) -> 'SimpleChoice':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SimpleChoice of this SimpleChoice.  # noqa: E501
        :rtype: SimpleChoice
        """
        return util.deserialize_model(dikt, cls)

    @property
    def identifier(self) -> str:
        """Gets the identifier of this SimpleChoice.


        :return: The identifier of this SimpleChoice.
        :rtype: str
        """
        return self._identifier

    @identifier.setter
    def identifier(self, identifier: str):
        """Sets the identifier of this SimpleChoice.


        :param identifier: The identifier of this SimpleChoice.
        :type identifier: str
        """
        if identifier is None:
            raise ValueError("Invalid value for `identifier`, must not be `None`")  # noqa: E501

        self._identifier = identifier

    @property
    def text(self) -> str:
        """Gets the text of this SimpleChoice.


        :return: The text of this SimpleChoice.
        :rtype: str
        """
        return self._text

    @text.setter
    def text(self, text: str):
        """Sets the text of this SimpleChoice.


        :param text: The text of this SimpleChoice.
        :type text: str
        """
        if text is None:
            raise ValueError("Invalid value for `text`, must not be `None`")  # noqa: E501

        self._text = text